+++ src/index.js
@@ -6,9 +6,10 @@ const api = require('./api.js');
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
 const cookieParser = require('cookie-parser');
+const api = require('./routes/index.js');
 
 const { logInfo } = require('./logger.js');
-const { PORT, API_BASE_PATH } = config;
+const { PORT } = config;
 
 // Body-parser middleware for JSON requests
 app.use(bodyParser.json());
@@ -19,8 +20,9 @@ app.use(cookieParser());
 // Logging
 app.use(morgan('tiny'));
 
-// Base API path default: /api/v1
-app.use(API_BASE_PATH, api);
+// Add routes
+app.use('/', api);
+
 app.listen(PORT, () => {
   logInfo(`API started on ${PORT}`);
 });
